Add a new API glamor_create_gc.
As at ValidateGC stage, it may need to touch the pixmap directly, for
example the tile pixmap. We must export this interface to DDX driver
and let the DDX driver to route the processing to us. As this pixmap
may be a texture only pixmap, and DDX don't know how to handle it.

Signed-off-by: Zhigang Gong <zhigang.gong@linux.intel.com>