glamor_color_convert: Let the caller to provide destination buffer.
As we don't need to allocate new buffer when downloading pixmap
to CPU, we change the prototype of the color converting function
and let the caller to provide the buffer to hold the result.

All the color conversion function supports store the result
just at the same place of the source.

Signed-off-by: Zhigang Gong <zhigang.gong@linux.intel.com>